requirements.txt
jupyter >=1.0.0
matplotlib >=3.6.3
numpy >=1.24.1
openai >=0.27.2
pandas >=1.5.3
scikit-learn >=1.2.0
seaborn >=0.12.2
torch >=2.0.0
torchaudio >=2.0.1
torchtext >=0.15.1
torchvision >=0.14.1
olive-oil-ml >=0.0.1
timm
openml
git+https://github.com/activatedgeek/tight-pac-bayes.git
